+++ properties/Connection/ContentIsAccessibleAfterCommit.php
@@ -5,9 +5,11 @@ namespace Properties\Formal\AccessLayer\Connection;
 
 use Formal\AccessLayer\{
     Query\SQL,
-    Query\Parameter,
+    Query\Insert,
     Query\StartTransaction,
     Query\Commit,
+    Table\Name,
+    Row,
     Connection,
 };
 use Innmind\BlackBox\{
@@ -51,15 +53,18 @@ final class ContentIsAccessibleAfterCommit implements Property
     {
         $connection(new StartTransaction);
 
-        $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
-            ->with(Parameter::of($this->uuid))
-            ->with(Parameter::of($this->username))
-            ->with(Parameter::of($this->number));
-        $connection($insert);
+        $connection(Insert::into(
+            new Name('test'),
+            Row::of([
+                'id' => $this->uuid,
+                'username' => $this->username,
+                'registerNumber' => $this->number,
+            ]),
+        ));
 
         $connection(new Commit);
 
-        $rows = $connection(SQL::of("SELECT * FROM `test` WHERE `id` = '{$this->uuid}'"));
+        $rows = $connection(SQL::of("SELECT * FROM test WHERE id = '{$this->uuid}'"));
 
         $assert->count(1, $rows);
         $assert->same(