+++ properties/Connection/SelectAliasedColumns.php
@@ -4,11 +4,11 @@ declare(strict_types = 1);
 namespace Properties\Formal\AccessLayer\Connection;
 
 use Formal\AccessLayer\{
-    Query\SQL,
-    Query\Parameter,
+    Query\Insert,
     Query\Select,
     Table\Name,
     Table\Column,
+    Row,
     Connection,
 };
 use Innmind\BlackBox\{
@@ -50,11 +50,14 @@ final class SelectAliasedColumns implements Property
 
     public function ensureHeldBy(Assert $assert, object $connection): object
     {
-        $insert = SQL::of('INSERT INTO `test` VALUES (?, ?, ?);')
-            ->with(Parameter::of($this->uuid))
-            ->with(Parameter::of($this->username))
-            ->with(Parameter::of($this->number));
-        $connection($insert);
+        $connection(Insert::into(
+            new Name('test'),
+            Row::of([
+                'id' => $this->uuid,
+                'username' => $this->username,
+                'registerNumber' => $this->number,
+            ]),
+        ));
 
         $select = Select::from($table = Name::of('test')->as('_test'))->columns(
             Column\Name::of('id')->in($table)->as('test_id'),